pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project]
name = "heuristic_evolution"
dynamic = [
  "version",
]

description = "Contextual Evolution of heuristics"
readme = "README.md"
requires-python = ">=3.9"

classifiers = [
  "Intended Audience :: Science/Research",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
  "absl-py",
  "click",
  "llm",
  "python-dotenv",
  "input-parser",
  "scipy",
  "numpy",
  "joblib",
  "plotly",
  "numba",
  "rl4co",
  "ortools",
  "pyvrp",
  "autopep8",
  "torch",
  "tiktoken",
  "pytest",
  "scholarly",
  "fuzzywuzzy",
  "scholarly"
]

[project.scripts]
ceoh = "solver.ceoh.__main__:main"
idea = "solver.ceoh.develope_ideas.__main__:main"
